Checkpoint, DI refactor
