+++ app/src/main/java/com/bluehub/fastmixer/common/dependencyinjection/Scope.kt
@@ -6,4 +6,7 @@ import javax.inject.Scope
 annotation class ApplicationScope
 
 @Scope
+annotation class ActivityScope
+
+@Scope
 annotation class FragmentScope
