+++ app/src/main/java/com/bluehub/fastmixer/common/dependencyinjection/components/ApplicationComponent.kt
@@ -3,8 +3,8 @@ package com.bluehub.fastmixer.common.dependencyinjection.components
 import android.app.Application
 import com.bluehub.fastmixer.MixerApplication
 import com.bluehub.fastmixer.common.dependencyinjection.ApplicationScope
-import com.bluehub.fastmixer.common.dependencyinjection.modules.application.ApplicationModule
-import com.bluehub.fastmixer.common.dependencyinjection.modules.presentation.PresentationModule
+import com.bluehub.fastmixer.common.dependencyinjection.modules.ActivityModule
+import com.bluehub.fastmixer.common.dependencyinjection.modules.AppModule
 import dagger.BindsInstance
 import dagger.Component
 import dagger.android.AndroidInjectionModule
@@ -13,8 +13,8 @@ import dagger.android.AndroidInjector
 @ApplicationScope
 @Component(modules = [
     AndroidInjectionModule::class,
-    ApplicationModule::class,
-    PresentationModule::class
+    AppModule::class,
+    ActivityModule::class
 ])
 interface ApplicationComponent: AndroidInjector<MixerApplication> {
     @Component.Factory
