+++ app/src/main/java/com/bluehub/fastmixer/common/dependencyinjection/modules/AppModule.kt
@@ -1,0 +1,14 @@
+package com.bluehub.fastmixer.common.dependencyinjection.modules
+
+import android.app.Application
+import android.content.Context
+import com.bluehub.fastmixer.common.dependencyinjection.ApplicationScope
+import dagger.Binds
+import dagger.Module
+
+@Module
+interface AppModule {
+    @Binds
+    @ApplicationScope
+    fun provideContext(application: Application): Context
+}
