+++ app/src/main/java/com/bluehub/fastmixer/screens/mixing/MixingRepository.kt
@@ -4,7 +4,7 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
-class MixingRepository @Inject constructor (val mixingEngineProxy: MixingEngineProxy) {
+class MixingRepository @Inject constructor(val mixingEngineProxy: MixingEngineProxy) {
     fun createMixingEngine() {
         mixingEngineProxy.create()
     }
